% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas.R
\name{ctwas}
\alias{ctwas}
\title{ctwas: Function to perform SuSiE-based causal tissue-gene pair and direct causal variant estimation}
\usage{
ctwas(by, bXest, LD, Noutcome, L.causal = 10, pip.thres.cred = 0.5)
}
\arguments{
\item{by}{A vector of Z-scores of the marginal effects from the outcome GWAS.}

\item{bXest}{A matrix of direct effect estimates based on the Z-scores of tissue-gene pairs.}

\item{LD}{The LD matrix of variants.}

\item{Noutcome}{The sample size of the outcome GWAS.}

\item{L.causal}{The number of single effects used for tissue-gene pairs and direct causal variants. Default is 10.}

\item{pip.thres.cred}{The cumulative PIP threshold for variables in a credible set. Default is 0.5.}
}
\value{
A list containing:
\describe{
\item{theta}{The estimated effects for tissue-gene pairs, scaled by the outcome GWAS sample size.}
\item{gamma}{The estimated effects for direct causal variants, scaled by the outcome GWAS sample size.}
\item{theta.se}{Standard errors of the tissue-gene pair effects.}
\item{gamma.se}{Standard errors of the direct causal variant effects.}
\item{theta.pip}{Posterior inclusion probabilities (PIP) for tissue-gene pairs.}
\item{gamma.pip}{Posterior inclusion probabilities (PIP) for direct causal variants.}
\item{theta.pratt}{Pratt index for tissue-gene pairs.}
\item{gamma.pratt}{Pratt index for direct causal variants.}
\item{theta.cs}{Credible set indicators for tissue-gene pairs.}
\item{gamma.cs}{Credible set indicators for direct causal variants.}
\item{theta.cs.pip}{Cumulative PIP within credible sets for tissue-gene pairs.}
\item{gamma.cs.pip}{Cumulative PIP within credible sets for direct causal variants.}
\item{fit.causal}{The SuSiE fit object for the causal analysis.}
\item{cs.summary}{A summary of the credible sets obtained from the analysis.}
}
}
\description{
This function performs the SuSiE (Sum of Single Effects) algorithm to estimate the causal tissue-gene pairs and direct causal variants using a provided set of summary statistics and linkage disequilibrium (LD) matrix. It returns a list of estimates and their associated statistics, including credible sets and Pratt estimations.
}
