% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design_matrix.R
\name{make_design_matrix}
\alias{make_design_matrix}
\title{make_design_matrix: Constructs a Design Matrix of Z-scores for Gene-Tissue Pairs}
\usage{
make_design_matrix(df)
}
\arguments{
\item{df}{A `data.table` with three columns: `SNP`, `Variable` (representing tissue-gene pairs), and `Zscore`.}
}
\value{
A matrix where rows correspond to SNPs, columns correspond to tissue-gene pairs, and values are z-scores. Missing values (`NA`) indicate that the tissue-gene pair does not have a z-score for that SNP.
}
\description{
This function constructs a design matrix from a `data.table` where each row represents a SNP, each column represents a tissue-gene pair, and the values are z-scores. If a tissue-gene pair does not have a z-score for a SNP, the corresponding entry will be `NA`.
}
\examples{
# Example usage:
df <- data.table::data.table(SNP = c("rs1", "rs1", "rs2"),
                             Variable = c("Gene1_Tissue1", "Gene1_Tissue2", "Gene2_Tissue1"),
                             Zscore = c(2.5, -1.2, 3.8))
design_matrix <- Make_Design_Matrix(df)
print(design_matrix)
}
