% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_predixcan.R
\name{modified_predixcan}
\alias{modified_predixcan}
\title{modified_predixcan: Function to estimate causal effects using a modified PrediXcan approach with BIC-based tuning parameter selection}
\usage{
modified_predixcan(
  by,
  bxest,
  LD,
  tauvec = seq(3, 10, by = 1),
  rho.gamma = 1.5,
  max.iter = 15,
  max.eps = 0.005,
  ebic.factor = 2,
  normmax = 2,
  pleiotropy.rm = NULL
)
}
\arguments{
\item{by}{A vector of Z-scores of marginal effects from the outcome GWAS (same as in ctwas).}

\item{bxest}{A vector of direct xQTL effect estimates for a tissue-gene pair.}

\item{LD}{The LD matrix of variants (same as in ctwas).}

\item{tauvec}{A vector of tuning parameters used in penalizing the direct causal effect. Default is `seq(3,10,by=1)`.}

\item{rho.gamma}{A parameter set in the ADMM algorithm. Default is 1.5.}

\item{max.iter}{The maximum number of iterations for the ADMM algorithm. Default is 15.}

\item{max.eps}{The convergence tolerance for the ADMM algorithm. Default is 0.005.}

\item{ebic.factor}{The extended BIC factor for model selection. Default is 2.}

\item{pleiotropy.rm}{A vector of indices specifying which variants should not be considered as having direct causal effects.}
}
\value{
A list containing:
\item{theta}{The estimated effect size of the tissue-gene pair.}
\item{gamma}{The estimated effect sizes of the direct causal variants.}
\item{covtheta}{The covariance of the estimated effect size `theta`.}
\item{Bic}{The BIC values for each tuning parameter.}
\item{Btheta}{The estimated `theta` values for each tuning parameter.}
\item{Bgamma}{The estimated `gamma` values for each tuning parameter.}
\item{Eta}{The estimated linear predictor for the tissue-gene pair.}
}
\description{
This function implements a modified version of the PrediXcan method, using an alternating direction method of multipliers (ADMM) algorithm and BIC for selecting the optimal tuning parameter. It estimates the causal effects of tissue-gene pairs while accounting for direct causal variants.
}
