% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poet_shrinkage.R
\name{poet_shrinkage}
\alias{poet_shrinkage}
\title{poet_shrinkage: Function to perform covariance matrix shrinkage using the POET method}
\usage{
poet_shrinkage(
  LD,
  KMax = min(15, round(nrow(LD)/2)),
  lamvec = seq(0.025, 0.25, by = 0.025),
  minvalue = 0.001
)
}
\arguments{
\item{LD}{A covariance matrix of linkage disequilibrium (LD).}

\item{KMax}{The maximum number of latent factors used in POET. Default is `min(15, round(nrow(LD)/2))`.}

\item{lamvec}{A vector of candidate shrinkage parameters. Default is `seq(0.025, 0.25, by = 0.025)`.}

\item{minvalue}{The threshold for the minimum eigenvalues used in determining the optimal shrinkage parameter. Default is `1e-3`.}
}
\value{
A covariance matrix that has been shrunk using the POET method.
}
\description{
This function performs shrinkage of the covariance matrix using the POET (Principal Orthogonal complEment Thresholding) method, with an automatic selection of the number of latent factors and the optimal shrinkage parameter.
}
