% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_partition.R
\name{R2_partition}
\alias{R2_partition}
\title{R2_partition: Function to partition the Pratt index contributions of tissue-gene pairs, direct causal variants, and infinitesimal effects}
\usage{
R2_partition(y, eta.theta, eta.gamma, eta.upsilon = NULL, LD)
}
\arguments{
\item{y}{A vector of Z-scores of marginal effects from the outcome GWAS.}

\item{eta.theta}{A vector of the linear predictor of tissue-gene causal effects.}

\item{eta.gamma}{A vector of the linear predictor of direct causal effects.}

\item{eta.upsilon}{A vector of the linear predictor of infinitesimal effects. Default is `NULL`.}

\item{LD}{A LD matrix.}
}
\value{
A data frame containing:
\item{r1}{Total Pratt index of all effects.}
\item{r2}{Pratt index contribution of tissue-gene pairs.}
\item{r3}{Pratt index contribution of direct causal variants.}
\item{r4}{Pratt index contribution of infinitesimal effects.}
}
\description{
This function partitions the total Pratt index of all effects into contributions from tissue-gene pairs, direct causal variants, and infinitesimal effects, using  vectors of linear predictors and the outcome GWAS Z-scores.
}
