% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgfm.R
\name{tgfm}
\alias{tgfm}
\title{tgfm: Function to perform eQTL fine-mapping and causal tissue-gene pair estimation}
\usage{
tgfm(
  by,
  bX,
  LD,
  Nvec,
  L.eqtl = 5,
  L.causal = 10,
  pip.thres.cred = 0.5,
  eqtl.sampling.time = 100,
  causal.sampling.time = 100,
  eqtl.thres = 0.05
)
}
\arguments{
\item{by}{A vector of Z-scores of the marginal effects from the outcome GWAS.}

\item{bX}{A matrix of Z-scores of marginal eQTL effect estimates for tissue-gene pairs.}

\item{LD}{The LD matrix of variants.}

\item{Nvec}{A vector representing the sample sizes, with the first element for the outcome GWAS and the subsequent elements for eQTL studies.}

\item{L.eqtl}{The number of single effects used in the eQTL fine-mapping step. Default is 5.}

\item{L.causal}{The number of single effects used for tissue-gene pairs and direct causal variants. Default is 10.}

\item{pip.thres.cred}{The cumulative PIP threshold for variables in a credible set. Default is 0.5.}

\item{eqtl.sampling.time}{The number of resampling iterations for eQTL effect estimation. Default is 100.}

\item{causal.sampling.time}{The number of resampling iterations for causal effect estimation. Default is 100.}

\item{eqtl.thres}{A threshold for the minimum individual PIPs in each credible set during eQTL fine-mapping. This threshold is used to pre-remove variables with very low PIPs to avoid overly large credible sets. Default is 0.05.}
}
\value{
A list containing:
\item{theta}{The estimated effects for tissue-gene pairs, scaled by the outcome GWAS sample size.}
\item{gamma}{The estimated effects for direct causal variants, scaled by the outcome GWAS sample size.}
\item{theta.se}{Standard errors of the tissue-gene pair effects.}
\item{gamma.se}{Standard errors of the direct causal variant effects.}
\item{theta.pip}{Posterior inclusion probabilities (PIP) for tissue-gene pairs.}
\item{gamma.pip}{Posterior inclusion probabilities (PIP) for direct causal variants.}
\item{theta.pratt}{Pratt estimations for tissue-gene pairs.}
\item{gamma.pratt}{Pratt estimations for direct causal variants.}
\item{theta.cs}{Credible set indicators for tissue-gene pairs.}
\item{gamma.cs}{Credible set indicators for direct causal variants.}
\item{theta.cs.pip}{PIP within credible sets for tissue-gene pairs.}
\item{gamma.cs.pip}{PIP within credible sets for direct causal variants.}
\item{fit.causal}{The SuSiE fit object for the causal analysis.}
\item{cs.summary}{A summary of the credible sets obtained from the analysis.}
\item{pip.causal.samping}{The PIP resampling results from causal effect estimation.}
\item{estimate.causal.sampling}{The effect size resampling results from causal effect estimation.}
\item{bXest}{The matrix of resampled eQTL effects used in causal effect estimation.}
}
\description{
This function performs fine-mapping of eQTLs and estimates causal tissue-gene pairs using summary statistics, linkage disequilibrium (LD) matrix, and resampling methods. It returns a list of estimates and associated statistics, including credible sets, PIP resampling, and Pratt estimations.
}
