% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgvis.R
\name{tgvis}
\alias{tgvis}
\title{tgvis: Function to estimate and select the optimal number of single effects using profile-likelihood and BIC}
\usage{
tgvis(
  by,
  bXest,
  LD,
  Noutcome,
  L.causal.vec = c(1:6),
  max.iter = 50,
  max.eps = 0.001,
  inner.iter = 50,
  pip.thres.cred = 0.5,
  varinf.upper.boundary = 0.3,
  varinf.lower.boundary = 0.01,
  ebic.beta = 1,
  ebic.upsilon = 1,
  pip.min = 0.05,
  pv.thres = 0.05,
  pleiotropy.rm = NULL
)
}
\arguments{
\item{by}{A vector of Z-scores of the marginal effects from the outcome GWAS.}

\item{bXest}{A matrix of direct effect estimates based on the Z-scores of tissue-gene pairs.}

\item{LD}{The LD matrix of variants.}

\item{Noutcome}{The sample size of the outcome GWAS.}

\item{L.causal.vec}{A vector of candidate numbers of single effects used in BIC. Default is `c(1:6)`.}

\item{max.iter}{The maximum number of iterations for the profile-likelihood algorithm. Default is 50.}

\item{max.eps}{The convergence tolerance for the profile-likelihood algorithm. Default is 1e-3.}

\item{inner.iter}{The maximum number of iterations for `susie_rss` within the profile-likelihood algorithm. Default is 50.}

\item{pip.thres.cred}{The cumulative PIP threshold for variables in a credible set. Default is 0.5.}

\item{varinf.upper.boundary}{The upper boundary for the prior variance of infinitesimal effects, multiplied by var(y) to adapt to different locus variances. Default is 0.3.}

\item{varinf.lower.boundary}{The lower boundary for the prior variance of infinitesimal effects, not multiplied by var(y). Default is 0.01.}

\item{ebic.beta}{The extended BIC factor for causal effects of tissue-gene pairs and direct causal variants used in BIC computation. Default is 1.}

\item{ebic.upsilon}{The extended BIC factor for infinitesimal effects used in BIC computation. Default is 1.}

\item{pip.min}{The minimum PIP threshold for individual causal effects in the profile-likelihood. This is used to specify which tissue-gene pairs and direct causal variants to include in the score test of variance of infinitesimal effects. Default is 0.05.}

\item{pv.thres}{The p-value threshold for the score test. Default is 0.05.}

\item{pleiotropy.rm}{A vector of indices specifying which variants should not be considered as having direct causal effects.}
}
\value{
A list containing:
\item{theta}{The estimated effects for tissue-gene pairs, scaled by the outcome GWAS sample size.}
\item{gamma}{The estimated effects for direct causal variants, scaled by the outcome GWAS sample size.}
\item{theta.se}{Standard errors of the tissue-gene pair effects.}
\item{gamma.se}{Standard errors of the direct causal variant effects.}
\item{theta.pip}{Posterior inclusion probabilities (PIP) for tissue-gene pairs.}
\item{gamma.pip}{Posterior inclusion probabilities (PIP) for direct causal variants.}
\item{theta.pratt}{Pratt estimations for tissue-gene pairs.}
\item{gamma.pratt}{Pratt estimations for direct causal variants.}
\item{theta.cs}{Credible set indicators for tissue-gene pairs.}
\item{gamma.cs}{Credible set indicators for direct causal variants.}
\item{theta.cs.pip}{PIP within credible sets for tissue-gene pairs.}
\item{gamma.cs.pip}{PIP within credible sets for direct causal variants.}
\item{upsilon}{The estimated infinitesimal effects.}
\item{var.upsilon}{The estimated variance of infinitesimal effects.}
\item{fit.causal}{The SuSiE fit object for the causal analysis.}
\item{cs.summary}{A summary of the credible sets obtained from the analysis.}
\item{Bicvec}{A vector of BIC values for each candidate number of single effects.}
}
\description{
This function estimates the number of single effects in a locus by combining profile-likelihood methods and Bayesian Information Criterion (BIC) to optimize the model. It includes resampling for estimating standard errors and performs score tests for infinitesimal effects.
}
